add connection between UVvisDataConcentration and UVvisConcentrationDetection
